Implement MovieInfo Data Layer

+++ core-data/src/main/java/com/jik/core/data/repository/MovieRepository.kt
@@ -7,5 +7,5 @@ interface MovieRepository {
 
     suspend fun getPopularMovies(page: Int): Result<List<Movie>>
 
-    suspend fun getMovieInfo(id: Int): Result<MovieInfo>
+    suspend fun getMovieInfo(id: Long): Result<MovieInfo>
 }
